refactor: 공용 컴포넌트 파일 이동 및 import 수정
