+++ src/app/(route)/(community)/news/news-detail/[id]/page.tsx
@@ -12,7 +12,7 @@ import { NewsTalkToolbar } from "../../../_components/NewsTalkToolbar";
 import NewsPostItem from "../../_components/NewsPostItem";
 import Copy from "@/app/_components/icon/Copy";
 import CommentBar from "@/app/_components/_gnb/_components/CommentBar";
-import SendCommentBox from "./_components/SendCommentBox";
+import SendCommentBox from "../../../_components/SendCommentBox";
 import { NewsItemType } from "@/app/_constants/newsItemType";
 import useGetNewsDataList from "@/_hooks/useGetNewsDataList";
 import useGetNewsInfoData from "@/_hooks/useGetNewsInfoData";
@@ -152,7 +152,7 @@ const Page = ({ params }: { params: Promise<{ id: string }> }) => {
       <div>
         <NewsTalkToolbar />
       </div>
-      <div className="w-[720px] min-h-[348px] rounded-[5px]">
+      <div className="w-[720px] min-h-[348px] rounded-b-[5px] rounded-bl-[5px] overflow-hidden">
         {sliceNewsListData?.map((data: NewsItemType) => (
           <div key={data.id}>
             <NewsPostItem newsItem={data} />
