+++ src/app/(route)/(community)/news/page.tsx
@@ -13,7 +13,7 @@ const Page = () => {
   console.log("NewsListData: ", data);
 
   return (
-    <div className="w-[720px] min-h-[1508px] flex flex-col justify-start bg-[#FAFAFA] rounded-[5px] overflow-hidden">
+    <div className="w-[720px] h-auto flex flex-col justify-start bg-[#FAFAFA] rounded-[5px] overflow-hidden">
       <div className="w-[720px] min-h-[120px] rounded-tl-[5px] rounded-tr-[5px] bg-[#FFFFFF] mx-auto">
         <NewsTalkToolbar />
       </div>
